add rating to course, when it is fetched from api show rating in course items

+++ app/src/main/java/org/stepic/droid/web/ApiImpl.java
@@ -752,7 +752,7 @@ public class ApiImpl implements Api {
     }
 
     @Override
-    public Single<CourseReviewResponse> getCourseReviews(long[] courseIds) {
+    public Single<CourseReviewResponse> getCourseReviews(int[] courseIds) {
         return loggedService.getCourseReviews(courseIds);
     }
